refactor(kysely): put effect-sql common code internal
